requirements.txt
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
PyYAML==3.13
requests==2.23.0
rtmbot==0.4.1
six==1.14.0
slackclient==1.3.2
urllib3==1.25.8
wcwidth==0.1.8
websocket-client==0.54.0
Werkzeug==1.0.0
zipp==3.0.0
